Optimize adjacency matrix creation for large datasets
Added a conditional approach for creating the adjacency matrix: 'fromLogical' for small N and 'fromIdx' for large N. This improves memory efficiency and performance when handling large datasets by avoiding the creation of large logical matrices. It is still not impossible to run out of memory, but much less likely now.